refactor(examples): splitting src of HitTestAPI example; split
